Place EB classes in EB common module
