+++ bin/deploy
@@ -55,16 +55,16 @@ app_bucket = apps.detect { |app| app.key == app_names[app_index] }
 application.bucket = app_bucket
 puts "App \"#{application.name}\" selected."
 
-require_relative '../deploy/eb_configuration'
-eb = EBConfiguration.new(app_bucket)
 require_relative '../deploy/s3_configuration'
 s3 = S3Configuration.new(app_bucket)
+require_relative '../deploy/eb/configuration'
+eb = Eb::Configuration.new(app_bucket)
 
-require_relative '../deploy/deployable/eb_platform'
 require_relative '../deploy/deployable/s3_platform'
+require_relative '../deploy/eb/platform'
 # TODO: Don't call out to bash, just call local classes to deploy
 if eb.exists?
-  application.platform = EBPlatform.new
+  application.platform = Eb::Platform.new
   application.platform_config = eb
   puts "Environment \'#{application.name}\' found on EB."
   puts "You've asked to deploy version \'#{TAG}\'."
